Storage tests
