+++ spring-cloud-gcp-autoconfigure/src/test/java/org/springframework/cloud/gcp/autoconfigure/sql/GcpCloudSqlAutoConfigurationTests.java
@@ -50,8 +50,9 @@ public class GcpCloudSqlAutoConfigurationTests {
 	private ApplicationContextRunner contextRunner = new ApplicationContextRunner()
 			.withPropertyValues("spring.cloud.gcp.sql.databaseName=test-database")
 			.withConfiguration(AutoConfigurations.of(GcpCloudSqlAutoConfiguration.class,
-					GcpContextAutoConfiguration.class, GcpCloudSqlTestConfiguration.class,
-					DataSourceAutoConfiguration.class));
+					GcpContextAutoConfiguration.class,
+					DataSourceAutoConfiguration.class))
+			.withUserConfiguration(GcpCloudSqlTestConfiguration.class);
 
 	@Test
 	public void testCloudSqlDataSourceTest() {
