+++ spring-cloud-gcp-dependencies/pom.xml
@@ -139,7 +139,7 @@
 			<dependency>
 				<groupId>com.google.cloud.sql</groupId>
 				<artifactId>mysql-socket-factory</artifactId>
-				<version>1.0.4</version>
+				<version>1.0.5</version>
 				<exclusions>
 					<exclusion>
 						<groupId>com.google.guava</groupId>
@@ -151,7 +151,7 @@
 			<dependency>
 				<groupId>com.google.cloud.sql</groupId>
 				<artifactId>postgres-socket-factory</artifactId>
-				<version>1.0.4</version>
+				<version>1.0.5</version>
 				<exclusions>
 					<exclusion>
 						<groupId>com.google.guava</groupId>
